+++ NEWS.md
@@ -1,5 +1,6 @@
 # rextendr (development version)
 
+* `use_extendr()` gets a new ability to overwrite existing rextendr templates (#292).
 * `use_extendr()` sets `publish = false` in the `[package]` section of the `Cargo.toml` (#297).
 
 # rextend 0.3.1